+++ kt_label_printing/models/__init__.py
@@ -1,1 +1,7 @@
-from . import kt_label_grid_utils, kt_label_png_export_mixin, kt_label_size
+# -*- coding: utf-8 -*-
+from . import (
+    kt_label_grid_utils,
+    kt_label_png_export_mixin,
+    kt_label_size,
+    product_label_layout,
+)